working code refactorization of the chat routes & scripts
