+++ project/routes/events-chat.routes.js
@@ -1,18 +1,141 @@
-const express = require("express");
-const router = express.Router();
-const Event = require("../models/Events.model")
-const Message = require("../models/Message.model")
 const {
   isEventMember
 } = require("../middlewares/event-guard");
 
+const express = require("express");
+const router = express.Router();
+const Event = require("../models/Events.model")
+const Message = require("../models/Message.model")
+const app = require('express')();
+const http = require('http').Server(app);
+const cors = require('cors');
+const io = require("socket.io")(http, {
+  cors: {
+    origin: "http://localhost:3000",
+    methods: ["GET", "POST"]
+  }
+});;
 
-router.post("/send-event-message/:eventId", async(req, res, next) => {
-   const {eventId} = req.params
-   const message = await Message.create(req.body)
-   await Event.findByIdAndUpdate(eventId, {$push:{messages: message}})
-   res.redirect(`/event-details/${eventId}`)
+// live group chat routes
 
+router.use(cors()); // IMPORTANT as we are working with 2 servers
+// most explorer will have problem with the Cros Cross-Origin Resource Sharing
+// so we have to configure it so it can receive data from our other port
+
+// see group chat and notifications
+router.get("/event-details/:eventId", [cors(), isEventMember], async (req, res, next) => {
+  const eventId = req.params.eventId
+  const event = await Event.findById(eventId)
+    .populate("joinRequests")
+    .populate({
+      path: "messages",
+      select: ["content", "author", "createdAt"],
+      populate: {
+        path: "author",
+        select: ["username", "_id"],
+      },
+      options: {
+        sort: { createdAt: 1 },
+      },
+    })
+  event.messages.forEach((message) => {
+    message.isYou = message.author._id == req.session.currentUser._id
+  })
+  const isAdmin = event.admin.includes(req.session.currentUser._id)
+  let notifications
+  let joinRequests
+  if (isAdmin) {
+    notifications = event.joinRequests.length
+  }
+  if (notifications) {
+    joinRequests = event.joinRequests
+    // console.log("esto tendría que ser sole el join request", joinRequests
+  }
+  res.render("events/event-details",
+    { session: req.session, event, isAdmin, notifications, joinRequests });
 });
+
+// send a message
+router.post("/event-details/:eventId",  cors(), async (req, res, next) => {
+    const eventId = req.params.eventId;
+    const message = await Message.create(req.body);
+
+    await Event.findByIdAndUpdate(eventId, { $push: { messages: message } });
+    io.emit("new message", message); // emit a new message event
+    res.redirect(`/event-details/${eventId}`);
+});
+
+// routes for group admins
+
+// DELETE EVENT
+router.post("/delete-event/:eventId", async(req, res, next) => {
+    const eventId = req.params.eventId
+    await Event.findByIdAndDelete(eventId)
+    res.redirect("/home")
+});
+
+//EDIT EVENT
+router.get("/edit-event/:eventId", isEventMember, async(req, res, next) => {
+    const eventId = req.params.eventId
+    const userId = req.session.currentUser._id
+    const event = await Event.findById(eventId).populate("members")
+    event.members.forEach((member)=>{
+            member.isAdmin = event.admin.includes(member._id)
+            member.isYou = member._id == userId
+    })
+    
+    res.render("events/edit", {session: req.session, event})
+});
+
+router.post("/edit-event/:eventId", async(req, res, next) => {
+    const eventId = req.params.eventId
+    const {name, description} = req.body
+    await Event.findByIdAndUpdate(eventId, {name, description})
+    
+    res.redirect(`/show-event/${eventId}`)
+});
+
+// MAKE GROUP MEMBER NEW ADMIN
+router.post("/make-admin/:newAdminId", async(req, res, next) => {
+    const newAdminId = req.params.newAdminId
+    const {eventId} = req.body
+    await Event.findByIdAndUpdate(eventId, {$push: { admin: newAdminId}})
+    
+    res.redirect(`/show-event/${eventId}`)
+});
+
+// KICK SOMEONE OUT OF THE GROUP
+router.post("/kick-out/:oldMember", async(req, res, next) => {
+    const oldMember = req.params.oldMember
+    const {eventId} = req.body
+    await Event.findByIdAndUpdate(eventId, {$pull: { members: oldMember}})
+    
+    res.redirect(`/show-event/${eventId}`)
+});
+
+
+//accept and reject notifications
+
+router.post("/decline-request", async(req, res, next) => {
+    const {askerId, eventId} = req.body
+    await Event.findByIdAndUpdate(eventId, {$pull: { joinRequests: askerId}})
+    res.redirect(`/event-details/${eventId}`)
+});
+
+router.post("/accept-request", async(req, res, next) => {
+    const {askerId, eventId} = req.body
+    await Event.findByIdAndUpdate(eventId, {$pull: { joinRequests: askerId}})
+    await Event.findByIdAndUpdate(eventId, {$push: { members: askerId}})
+    res.redirect(`/event-details/${eventId}`)
+})
+
+// CONNECT TO THE SOCKET.IO SERVER
+
+http.listen(4000, () => {
+  console.log("socket.io listening on *:4000");
+});
+
+
+
 
 module.exports = router;
